move `Transition` to its own file
