+++ examples/control_vector.py
@@ -3,7 +3,7 @@ from lukefi.metsi.domain.events import DoNothing
 from lukefi.metsi.domain.forestry_types import ForestCondition
 from lukefi.metsi.domain.natural_processes.grow_acta import grow_acta_fn
 from lukefi.metsi.domain.pre_ops import filter_stands, filter_trees, generate_reference_trees
-from lukefi.metsi.sim.sim_configuration import Transition
+from lukefi.metsi.sim.transition import Transition
 from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
 from lukefi.metsi.sim.generators import Sequence
 from examples.declarations.sqlite import sqlite_decl
@@ -31,7 +31,7 @@ control_structure = {
         ],
         filter_stands: [
             {
-                "remove": lambda stand: (stand.site_type_category is None) or (stand.site_type_category == 0) 
+                "remove": lambda stand: (stand.site_type_category is None) or (stand.site_type_category == 0)
             }
         ],
         filter_trees: [
