+++ tests/robot/VMI13_mini_db/input/control.py
@@ -5,7 +5,7 @@ from lukefi.metsi.domain.pre_ops import filter_stands, filter_trees, generate_re
 from lukefi.metsi.sim.condition import Condition
 from lukefi.metsi.sim.generators import Alternatives
 from lukefi.metsi.sim.operations import do_nothing
-from lukefi.metsi.sim.sim_configuration import Transition
+from lukefi.metsi.sim.transition import Transition
 from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
